+++ package.json
@@ -66,11 +66,11 @@
     ]
   },
   "dependencies": {
+    "@babel/polyfill": "^7.0.0-beta.44",
     "@digix/tempo": "^0.2.0",
     "@gnosis.pm/dutch-exchange-smartcontracts": "^0.3.7",
     "autobind-decorator": "^2.1.0",
     "await-delay": "^1.0.0",
-    "babel-polyfill": "^6.26.0",
     "bignumber.js": "^5.0.0",
     "blockies": "0.0.2",
     "classnames": "^2.2.5",
@@ -88,7 +88,7 @@
     "react-custom-scroll": "2.0.1",
     "react-datetime": "^2.10.1",
     "react-dom": "^16.2.0",
-    "react-hot-loader": "^3.0.0",
+    "react-hot-loader": "^4.0.1",
     "react-intl": "2.4.0",
     "react-modal": "^3.0.0",
     "react-portal": "^3.0.0",
